+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-Version 0.1.3
+Version 0.1.4
+------------------------
+0a01ae0 - fix detekt issues (Sat, 16 Jun 2018 19:03:15 +0100) <Elliot Tormey>
+082bddb - Outage list item #10 (Sat, 16 Jun 2018 18:57:37 +0100) <Elliot Tormey>
+3b8287f - remove some missing outageDetail objects from resources/outages.json (Sat, 16 Jun 2018 17:52:35 +0100) <Elliot Tormey>
+8ae63fb - Outage location #9 (Sat, 16 Jun 2018 17:52:14 +0100) <Elliot Tormey>
+ba3acf7 - add swipe refresh widget to OutageLit (Sat, 16 Jun 2018 16:45:10 +0100) <Elliot Tormey>
+a2ae350 - convert OutageMap to Controller (Sat, 16 Jun 2018 16:34:40 +0100) <Elliot Tormey>
+7cdd475 - add info item to bottom bar (Sat, 16 Jun 2018 16:34:30 +0100) <Elliot Tormey>
+0c232c9 - convert OutageList to Controller, create MainActivity as controller main router, add bottomNav to MainActivity (Sat, 16 Jun 2018 16:00:31 +0100) <Elliot Tormey>
+
+Version 0.1.3 - e2d8971
 ------------------------
 8538d4f - add threetenbp to handle date things (Sat, 16 Jun 2018 15:09:01 +0100) <Elliot Tormey>
 5467083 - update circleci/config to store all reports in artifacts (Sat, 16 Jun 2018 14:31:05 +0100) <Elliot Tormey>